Busca ativa, e a primeira tela que abre sem conta
O edital dá ao perfil Público "pesquisar profissionais (especialidade,
experiência, nome); ver perfil" (Anexo I, item 3), e até aqui nenhuma
rota atendia quem não tinha sessão. Agora /profissionais e /perfil/{id}
atendem. O anônimo não vê menos tela, vê menos dado:
Visibilidade::alcanceDe() entrega a ele só o que cada titular marcou
como PUBLICO, e quem entrou alcança também o de AUTENTICADO. Medido num
perfil reservado: anônimo vê 0 campos e 0 ARTs, logado vê 4 campos e 2
ARTs.

Corrige um defeito que a abertura da rota criou e que nunca chegou a
existir em produção: publico() fazia (int) Sessao::usuarioId(), e para
quem não entrou isso vira 0, que não é null. A Visao decide `autenticado`
por !== null, então o anônimo teria recebido o alcance de quem tem conta.

A busca é filtro, não motor. Não há score, limiar nem semente, e é por
isso que a ordem alfabética não vira ranking (item 10.1): o termo decide
quem entra, e dentro de quem entrou ninguém está na frente de ninguém.
O teto de 60 existe menos pela tela e mais pelo item 10.4 — busca sem
teto aberta a anônimo é um endpoint que entrega a massa inteira numa
requisição. Truncamento é avisado, nunca silencioso.

O portão global é o mesmo `perfisAbertos()` do motor: candidato que não
entra no feed não pode ser encontrado por outro caminho. Campo a campo,
a mesma Visao que monta o perfil — o resumo de quem abriu só para
autenticados não vaza no resultado de um anônimo.

Campo nulo não vira "Não informado" na tela. Ele tem dois sentidos que a
interface não distingue, "não preencheu" e "não abriu para você", e
afirmar o primeiro seria escrever algo falso sobre a pessoa; o bloco
simplesmente não é renderizado.

O resultado também explica por que casou. A busca procura no subgrupo da
Tabela de Obras e Serviços, mas o cartão só mostrava o grupo: quem
digitava "elétrica" recebia quatro pessoas cujo grupo diz
"Eletrotécnica", certo e inexplicável ao mesmo tempo. Agora o subgrupo
que casa vem na frente, e a modalidade do registro entra no cartão —
ela é a leitura mais literal de "especialidade" no Anexo I e a busca já
casava nela sem nunca mostrá-la.

42 conferências de padrão com 0 violações · 27 telas · 188 testes ·
E4 32 · E2 138.

+++ src/Controller/PerfilController.php
@@ -65,8 +65,13 @@ final class PerfilController
      */
     public function publico(string $id): string
     {
-        $alvoId     = (int) $id;
-        $espectador = (int) Sessao::usuarioId();
+        $alvoId = (int) $id;
+
+        // Null quando ninguém entrou, e **não** zero: a Visao decide `autenticado` por
+        // `!== null`, então um 0 aqui daria ao anônimo o alcance de quem tem conta, entregando
+        // o que cada titular abriu só para autenticados. A rota é pública (Anexo I, item 3), e é
+        // exatamente por isso que a distinção precisa sobreviver até aqui.
+        $espectador = Sessao::usuarioId();
 
         // Conta inexistente, excluída ou bloqueada não volta de perfisAtivos, e a ausência fecha
         // o perfil — mesmo efeito da exclusão lógica (D05) e do bloqueio da E6.
